requirements.txt
dj-config-url==0.1.1
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
mysqlclient==1.3.13
pytz==2018.5
whitenoise==4.1
